+++ springql-core/src/pipeline/pump_model/window_operation_parameter.rs
@@ -3,11 +3,11 @@
 pub(crate) mod aggregate;
 pub(crate) mod join_parameter;
 
-use self::{aggregate::GroupAggregateParameter, join_parameter::JoinParameter};
+use self::{aggregate::AggregateParameter, join_parameter::JoinParameter};
 
 /// Window operation parameters
 #[derive(Clone, PartialEq, Debug)]
 pub(crate) enum WindowOperationParameter {
-    GroupAggregation(GroupAggregateParameter),
+    Aggregate(AggregateParameter),
     Join(JoinParameter),
 }