+++ springql-core/src/pipeline/pump_model/window_operation_parameter/aggregate.rs
@@ -2,21 +2,32 @@
 
 use crate::expr_resolver::expr_label::{AggrExprLabel, ValueExprLabel};
 
+/// [GROUP BY c1, c2, c3...]
+#[derive(Clone, PartialEq, Debug, Default, new)]
+pub(crate) struct GroupByLabels(
+    /// Empty when GROUP BY clause is not supplied.
+    Vec<ValueExprLabel>,
+);
+impl GroupByLabels {
+    pub(crate) fn as_labels(&self) -> &[ValueExprLabel] {
+        &self.0
+    }
+}
+
 /// TODO [support complex expression with aggregations](https://gh01.base.toyota-tokyo.tech/SpringQL-internal/SpringQL/issues/152)
 ///
 /// ```sql
 /// SELECT group_by, aggr_expr.func(aggr_expr.aggregated)
 ///   FROM s
-///   GROUP BY group_by
+///   [GROUP BY group_by]
 ///   SLIDING WINDOW ...;
 /// ```
-#[derive(Copy, Clone, PartialEq, Debug, new)]
-pub(crate) struct GroupAggregateParameter {
+#[derive(Clone, PartialEq, Debug, new)]
+pub(crate) struct AggregateParameter {
     // TODO multiple aggr_expr
     pub(crate) aggr_func: AggregateFunctionParameter,
     pub(crate) aggr_expr: AggrExprLabel,
-
-    pub(crate) group_by: ValueExprLabel, // TODO multiple group by expression
+    pub(crate) group_by: GroupByLabels,
 }
 
 #[derive(Copy, Clone, Eq, PartialEq, Debug)]